+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Change Log
 
-## [v1.0.1] (https://github.com/zingat/SwipeUpView/releases/tag/1.1.0)
+## [v1.2.0] (https://github.com/zingat/SwipeUpView/releases/tag/1.2.0)
+- These methods were added. 
+- - goToUp(stepCount : Int)
+- - goToTop()
+- - goToDown(stepCount : Int)
+- - goToBottom()
+
+## [v1.1.0] (https://github.com/zingat/SwipeUpView/releases/tag/1.1.0)
 - Adding static heights 
 
 ## [v1.0.1] (https://github.com/zingat/SwipeUpView/releases/tag/1.0.1)